CLOUDSTACK-10358: SSH keys are missing on Config Drive disk in some cases
