+++ test/integration/plugins/nuagevsp/test_nuage_configdrive.py
@@ -42,37 +42,16 @@ from marvin.lib.utils import random_gen
 from nose.plugins.attrib import attr
 
 # Import Local Modules
-from integration.plugins.nuagevsp.nuage_lib import MySSHKeyPair
-from nuageTestCase import nuageTestCase
+from component.test_configdrive import MySSHKeyPair, ConfigDriveUtils
+from nuageTestCase import nuageTestCase, needscleanup
 
 NO_SUCH_FILE = "No such file or directory"
 
 
-class TestNuageConfigDrive(nuageTestCase):
+class TestNuageConfigDrive(nuageTestCase, ConfigDriveUtils):
     """Test user data and password reset functionality
     using configDrive with Nuage VSP SDN plugin
     """
-
-    class CreateResult:
-        def __init__(self, success, offering=None, network=None, vpc=None):
-            self.success = success
-            self.network = network
-            self.offering = offering
-            self.vpc = vpc
-
-    class PasswordTest:
-        def __init__(self, password):
-            self.test_presence = False
-            self.presence = None
-            self.password = None
-            if type(password) is bool:
-                self.test_presence = True
-                self.presence = password
-                self.password = None
-            elif type(password) is unicode or type(password) is str:
-                self.test_presence = True
-                self.presence = True
-                self.password = password
 
     class StartVM(threading.Thread):
 
@@ -219,472 +198,11 @@ class TestNuageConfigDrive(nuageTestCase):
         self.updateTemplate(False)
         return
 
-    def updateTemplate(self, value):
-        """Updates value of the guest VM template's password enabled setting
-        """
-        self.debug("Updating value of guest VM template's password enabled "
-                   "setting")
-        cmd = updateTemplate.updateTemplateCmd()
-        cmd.id = self.template.id
-        cmd.passwordenabled = value
-        self.api_client.updateTemplate(cmd)
-        list_template_response = list_templates(self.api_client,
-                                                templatefilter="all",
-                                                id=self.template.id
-                                                )
-        self.template = list_template_response[0]
-        self.debug("Updated guest VM template")
-
-    def get_userdata_url(self, vm):
-        """Returns user data URL for the given VM object"""
-        self.debug("Getting user data url")
-        nic = vm.nic[0]
-        gateway = str(nic.gateway)
-        self.debug("Gateway: " + gateway)
-        user_data_url = 'curl "http://' + gateway + ':80/latest/user-data"'
-        return user_data_url
-
-    def create_and_verify_fip_and_fw(self, vm, public_ip, network):
-        """
-        Creates and verifies (Ingress) firewall rule
-        with a Static NAT rule enabled public IP"""
-        self.debug("Creating and verifying firewall rule")
-        self.create_StaticNatRule_For_VM(vm, public_ip, network)
-
-        # VSD verification
-        self.verify_vsd_floating_ip(network, vm, public_ip.ipaddress)
-
-        fw_rule = self.create_FirewallRule(
-            public_ip, self.test_data["ingress_rule"])
-
-        # VSD verification
+    def validate_firewall_rule(self, fw_rule):
         self.verify_vsd_firewall_rule(fw_rule)
-        self.debug("Successfully created and verified firewall rule")
 
-    def mount_config_drive(self, ssh):
-        """
-        This method is to verify whether configdrive iso
-        is attached to vm or not
-        Returns mount path if config drive is attached else False
-        """
-        mountdir = "/root/iso"
-        cmd = "blkid -t LABEL='config-2' /dev/sr? /dev/hd? /dev/sd? /dev/xvd? -o device"
-        tmp_cmd = [
-            'bash -c "if [ ! -d /root/iso ] ; then mkdir /root/iso ; fi"',
-            "umount /root/iso"]
-        for tcmd in tmp_cmd:
-            ssh.execute(tcmd)
-        configDrive = ssh.execute(cmd)
-        res = ssh.execute("mount {} {}".format(str(configDrive[0]), mountdir))
-        if str(res).lower().find("mounting read-only") > -1:
-            self.debug("configDrive iso is mounted at location %s" % mountdir)
-            return mountdir
-        else:
-            return None
-
-    def _get_config_drive_data(self, ssh, file, name, fail_on_missing=True):
-        """Fetches the content of a file file on the config drive
-
-        :param ssh: SSH connection to the VM
-        :param file: path to the file to fetch
-        :param name: description of the file
-        :param fail_on_missing:
-                 whether the test should fail if the file is missing
-        :type ssh: marvin.sshClient.SshClient
-        :type file: str
-        :type name: str
-        :type fail_on_missing: bool
-        :returns: the content of the file
-        :rtype: str
-        """
-        cmd = "cat %s" % file
-        res = ssh.execute(cmd)
-        content = '\n'.join(res)
-
-        if fail_on_missing and NO_SUCH_FILE in content:
-            self.debug("{} is not found".format(name))
-            self.fail("{} is not found".format(name))
-
-        return content
-
-    def verify_config_drive_data(self, ssh, file, expected_content, name):
-        """Verifies that the file contains the expected content
-
-        :param ssh: SSH connection to the VM
-        :param file: path to the file to verify
-        :param expected_content:
-        :param name:
-        :type ssh: marvin.sshClient.SshClient
-        :type file: str
-        :type expected_content: str
-        :type name: str
-        """
-        actual_content = self._get_config_drive_data(ssh, file, name)
-
-        self.debug("Expected {}: {}".format(name, expected_content))
-        self.debug("Actual {}: {}".format(name, actual_content))
-
-        self.assertEqual(expected_content, actual_content,
-                         'Userdata found: %s is not equal to expected: %s'
-                         % (actual_content, expected_content))
-
-    def verifyUserData(self, ssh, iso_path, userdata):
-        """
-        verify Userdata
-
-        :param ssh: SSH connection to the VM
-        :param iso_path: mount point of the config drive
-        :param userdata: Expected userdata
-        :type ssh: marvin.sshClient.SshClient
-        :type iso_path: str
-        :type userdata: str
-        """
-        self.verify_config_drive_data(
-            ssh,
-            iso_path + "/cloudstack/userdata/user_data.txt",
-            userdata,
-            "userdata (ACS)"
-        )
-
-    def verifyOpenStackUserData(self, ssh, iso_path, userdata):
-        """
-        verify Userdata in Openstack format
-
-        :param ssh: SSH connection to the VM
-        :param iso_path: mount point of the config drive
-        :param userdata: Expected userdata
-        :type ssh: marvin.sshClient.SshClient
-        :type iso_path: str
-        :type userdata: str
-        """
-        self.verify_config_drive_data(
-            ssh,
-            iso_path + "/openstack/latest/user_data",
-            userdata,
-            "userdata (Openstack)"
-        )
-
-    def verifyPassword(self, ssh, iso_path, password):
-        self.debug("Expected VM password is %s " % password.password)
-        password_file = iso_path + "/cloudstack/password/vm_password.txt"
-        vmpassword = self._get_config_drive_data(ssh, password_file,
-                                                 "ConfigDrive password",
-                                                 fail_on_missing=False)
-
-        self.debug("ConfigDrive password is %s " % vmpassword)
-
-        if NO_SUCH_FILE in vmpassword:
-            self.debug("Password file is not found")
-            return False, False
-        elif (password.password is not None) \
-                and (password.password in vmpassword):
-            self.debug("Expected Password is found in configDriveIso")
-            return True, True
-        else:
-            self.debug("Expected password is not found in configDriveIso")
-            return True, False
-
-    def verifySshKey(self, ssh, iso_path, ssh_key):
-        self.debug("Expected VM sshkey is %s " % ssh_key.name)
-        publicKey_file = iso_path + "/cloudstack/metadata/public-keys.txt"
-        cmd = "ssh-keygen -lf %s | cut -f2 -d' '" % publicKey_file
-        res = ssh.execute(cmd)
-        vmsshkey = str(res[0])
-
-        self.debug("ConfigDrive ssh key is %s " % vmsshkey)
-
-        if NO_SUCH_FILE in vmsshkey:
-            self.fail("SSH keyfile is not found")
-
-        self.assertEqual(
-            vmsshkey,
-            ssh_key.fingerprint,
-            "Fingerprint of authorized key does not match ssh key fingerprint"
-        )
-
-    def verifyMetaData(self, vm, ssh, iso_path):
-        """
-        verify metadata files in CloudStack format
-
-        :param vm: the VM
-        :param ssh: SSH connection to the VM
-        :param iso_path: mount point of the config drive
-        :type vm: VirtualMachine
-        :type ssh: marvin.sshClient.SshClient
-        :type iso_path: str
-        """
-
-        metadata_dir = iso_path + "/cloudstack/metadata/"
-        vm_files = ["availability-zone.txt",
-                    "service-offering.txt",
-                    "instance-id.txt",
-                    "vm-id.txt",
-                    "local-hostname.txt",
-                    "local-ipv4.txt",
-                    "public-ipv4.txt"]
-
-        get_name = lambda file: \
-            "{} metadata".format(file.split('.'[-1].replace('-', ' ')))
-
-        metadata = {vm_file:
-                        self._get_config_drive_data(ssh,
-                                                    metadata_dir + vm_file,
-                                                    get_name(vm_file))
-                    for vm_file in vm_files}
-
-        self.assertEqual(
-            str(metadata["availability-zone.txt"]),
-            self.zone.name,
-            "Zone name inside metadata does not match with the zone"
-        )
-        self.assertEqual(
-            str(metadata["local-hostname.txt"]),
-            vm.instancename,
-            "vm name inside metadata does not match with the "
-            "instance name"
-        )
-        self.assertEqual(
-            str(metadata["vm-id.txt"]),
-            vm.id,
-            "vm name inside metadata does not match with the "
-            "instance name"
-        )
-        self.assertEqual(
-            str(metadata["instance-id.txt"]),
-            vm.id,
-            "vm name inside metadata does not match with the "
-            "instance name"
-        )
-        self.assertEqual(
-            str(metadata["service-offering.txt"]),
-            vm.serviceofferingname,
-            "Service offering inside metadata does not match "
-            "with the instance offering"
-        )
-        return
-
-    def verifyOpenStackData(self, ssh, iso_path):
-        """
-        verify existence of metadata and user data files in OpenStack format
-
-        :param ssh: SSH connection to the VM
-        :param iso_path: mount point of the config drive
-        :type ssh: marvin.sshClient.SshClient
-        :type iso_path: str
-        """
-        openstackdata_dir = iso_path + "/openstack/latest/"
-        openstackdata_files = ["meta_data.json",
-                               "vendor_data.json",
-                               "network_data.json"]
-        for file in openstackdata_files:
-            res = ssh.execute("cat %s" % openstackdata_dir + file)
-            if NO_SUCH_FILE in res[0]:
-                self.fail("{} file not found in vm openstack".format(file))
-
-    @needscleanup
-    def generate_ssh_keys(self):
-        """Generates ssh key pair
-
-        Writes the private key into a temp file and returns the file name
-
-        :returns: path of the private key file
-
-        """
-        keypair = MySSHKeyPair.create(
-            self.api_client,
-            name=random_gen() + ".pem",
-            account=self.account.user[0].account,
-            domainid=self.account.domainid)
-
-        self.debug("Created keypair with name: %s" % keypair.name)
-        self.debug("Writing the private key to local file")
-        keypair.private_key_file = tempfile.gettempdir() + os.sep + keypair.name
-        self.tmp_files.append(keypair.private_key_file)
-        self.debug("File path: %s" % keypair.private_key_file)
-        with open(keypair.private_key_file, "w+") as f:
-            f.write(keypair.privatekey)
-        os.chmod(keypair.private_key_file, 0o400)
-        return keypair
-
-    def umount_config_drive(self, ssh, iso_path):
-        """unmount config drive inside guest vm
-
-        :param ssh: SSH connection to the VM
-        :param iso_path: mount point of the config drive
-        :type ssh: marvin.sshClient.SshClient
-        :type iso_path: str
-        """
-        ssh.execute("umount -d %s" % iso_path)
-        # Give the VM time to unlock the iso device
-        time.sleep(2)
-        # Verify umount
-        result = ssh.execute("ls %s" % iso_path)
-        self.assertTrue(len(result) == 0,
-                        "After umount directory should be empty "
-                        "but contains: %s" % result)
-
-    def update_provider_state(self, new_state):
-        """
-        Enables or disables the ConfigDrive Service Provider
-
-        :param new_state: "Enabled" | "Disabled"
-        :type new_state: str
-        :return: original state
-        :rtype: str
-        """
-        self.debug("Updating Service Provider ConfigDrive to %s" % new_state)
-        configdriveprovider = NetworkServiceProvider.list(
-            self.api_client,
-            name="ConfigDrive",
-            physicalnetworkid=self.vsp_physical_network.id)[0]
-        orig_state = configdriveprovider.state
-        NetworkServiceProvider.update(self.api_client,
-                                      configdriveprovider.id,
-                                      state=new_state)
-        self.validate_NetworkServiceProvider("ConfigDrive", state=new_state)
-        return orig_state
-
-    def verify_network_creation(self, offering=None,
-                                offering_name=None,
-                                gateway=None,
-                                vpc=None, acl_list=None, testdata=None):
-        """
-        Creates a network
-
-        :param offering: Network Offering
-        :type offering: NetworkOffering
-        :param offering_name: Offering name
-        :type offering_name: Optional[str]
-        :param gateway: gateway
-        :type gateway: str
-        :param vpc: in case of a VPC tier, the parent VPC
-        :type vpc: VPC
-        :param acl_list: in case of a VPC tier, the acl list
-        :type acl_list: NetworkACLList
-        :param testdata: Test data
-        :type testdata: dict
-        :return: Network Creation Result
-        :rtype: CreateResult
-        """
-        if offering is None:
-            self.debug("Creating Nuage VSP network offering...")
-            offering = self.create_NetworkOffering(
-                self.test_data["nuagevsp"][offering_name])
-            self.validate_NetworkOffering(offering, state="Enabled")
-        try:
-            network = self.create_Network(offering,
-                                          gateway=gateway,
-                                          vpc=vpc,
-                                          acl_list=acl_list,
-                                          testdata=testdata)
-            return self.CreateResult(True, offering=offering, network=network)
-        except Exception:
-            self.debug("Exception: %s" % sys.exc_info()[0])
-            return self.CreateResult(False, offering=offering)
-
-    def verify_vpc_creation(self, offering=None, offering_name=None):
-        """
-        Creates a VPC
-
-        :param offering: VPC Offering
-        :type offering: VpcOffering
-        :param offering_name: Offering name
-        :type offering_name: Optional[str]
-        :return: VPC Creation Result
-        :rtype: CreateResult
-        """
-        if offering is None:
-            self.debug("Creating Nuage VSP VPC offering...")
-            offering = self.create_VpcOffering(
-                self.test_data["nuagevsp"][offering_name])
-            self.validate_VpcOffering(offering, state="Enabled")
-        try:
-            vpc = self.create_Vpc(offering, cidr='10.1.0.0/16')
-            self.validate_Vpc(vpc, state="Enabled")
-            return self.CreateResult(True, offering=offering, vpc=vpc)
-        except Exception:
-            return self.CreateResult(False, offering=offering)
-
-    def update_password_enable_in_template(self, new_state):
-        self.debug("Updating guest VM template to password %s" % new_state)
-        orig_state = self.template.passwordenabled
-        if self.template.passwordenabled is not new_state:
-            self.updateTemplate(new_state)
-        self.assertEqual(self.template.passwordenabled, new_state,
-                         "Guest VM template is not password enabled")
-        return orig_state
-
-    def verify_config_drive_content(self, vm,
-                                    public_ip,
-                                    password_test,
-                                    userdata=None,
-                                    metadata=False,
-                                    ssh_key=None,
-                                    ssh_client=None):
-        """Verify Config Drive Content
-
-        :param vm:
-        :param public_ip:
-        :param password_test:
-        :param userdata:
-        :param metadata:
-        :param sshkey:
-        :param ssh_client: SSH Connection
-        :type vm:
-        :type public_ip:
-        :type password_test:
-        :type userdata: object
-        :type metadata:
-        :type sshkey:
-        :type ssh_client:
-        :return: SSH Connection
-        """
-
-        if self.isSimulator:
-            self.debug(
-                "Simulator Environment: Skipping Config Drive content verification")
-            return
-
-        self.debug("SSHing into the VM %s" % vm.name)
-        if ssh_client is None:
-            ssh = self.ssh_into_VM(vm, public_ip)
-        else:
-            ssh = ssh_client
-        d = {x.name: x for x in ssh.logger.handlers}
-        ssh.logger.handlers = list(d.values())
-        config_drive_path = self.mount_config_drive(ssh)
-        self.assertIsNotNone(config_drive_path,
-                             'ConfigdriveIso is not attached to vm')
-        if metadata:
-            self.debug("Verifying metadata for vm: %s" % vm.name)
-            self.verifyMetaData(vm, ssh, config_drive_path)
-            self.debug("Verifying openstackdata for vm: %s" % vm.name)
-            self.verifyOpenStackData(ssh, config_drive_path)
-
-        if userdata is not None:
-            self.debug("Verifying userdata for vm: %s" % vm.name)
-            self.verifyUserData(ssh, config_drive_path, userdata)
-            self.verifyOpenStackUserData(ssh, config_drive_path, userdata)
-
-        if password_test.test_presence:
-            self.debug("Verifying password for vm: %s" % vm.name)
-            test_result = self.verifyPassword(ssh, config_drive_path,
-                                              password_test)
-            self.assertEqual(test_result[0], password_test.presence,
-                             "Expected is that password is present: %s "
-                             " but found is: %s"
-                             % (test_result[0], password_test.presence))
-
-        if password_test.password is not None:
-            self.debug("Password for vm is %s" % password_test.password)
-            self.assertEqual(test_result[1], True,
-                             "Password value test failed.")
-        if ssh_key is not None:
-            self.debug("Verifying sshkey for vm: %s" % vm.name)
-            self.verifySshKey(ssh, config_drive_path, ssh_key)
-
-        self.umount_config_drive(ssh, config_drive_path)
-        return ssh
+    def validate_StaticNat_rule_For_VM(self, public_ip, network, vm):
+        self.verify_vsd_floating_ip(network, vm, public_ip.ipaddress)
 
     def create_guest_vm(self, networks, acl_item=None,
                         vpc=None, keypair=None):
@@ -710,102 +228,6 @@ class TestNuageConfigDrive(nuageTestCase):
         if acl_item is not None:
             self.verify_vsd_firewall_rule(acl_item)
         return vm
-
-    # nic_operation_VM - Performs NIC operations such as add, remove, and
-    # update default NIC in the given VM and network
-    def nic_operation_VM(self, vm, network, operation="add"):
-        self.debug("Performing %s NIC operation in VM with ID - %s and "
-                   "network with ID - %s" % (operation, vm.id, network.id))
-        if operation is "add":
-            vm.add_nic(self.api_client, network.id)
-            self.debug("Added NIC in VM with ID - %s and network with ID - %s"
-                       % (vm.id, network.id))
-        vm_info = VirtualMachine.list(self.api_client, id=vm.id)[0]
-        for nic in vm_info.nic:
-            if nic.networkid == network.id:
-                nic_id = nic.id
-        if operation is "update":
-            vm.update_default_nic(self.api_client, nic_id)
-            self.debug("Updated default NIC to NIC with ID- %s in VM with ID "
-                       "- %s and network with ID - %s" %
-                       (nic_id, vm.id, network.id))
-        if operation is "remove":
-            vm.remove_nic(self.api_client, nic_id)
-            self.debug("Removed NIC with ID - %s in VM with ID - %s and "
-                       "network with ID - %s" % (nic_id, vm.id, network.id))
-
-    def update_userdata(self, vm, new_user_data):
-        """Updates the user data of a VM
-
-        :param vm: the Virtual Machine
-        :param new_user_data: UserData to set
-        :type vm: VirtualMachine
-        :type new_user_data: str
-        :returns: User data in base64 format
-        :rtype: str
-        """
-        updated_user_data = base64.b64encode(new_user_data)
-        vm.update(self.api_client, userdata=updated_user_data)
-        return new_user_data
-
-    def reset_password(self, vm):
-        vm.password = vm.resetPassword(self.api_client)
-        self.debug("Password reset to - %s" % vm.password)
-        self.debug("VM - %s password - %s !" %
-                   (vm.name, vm.password))
-
-    def wait_until_done(self, thread_list, name):
-        for aThread in thread_list:
-            self.debug("[Concurrency]Join %s for vm %s" % (name,
-                                                           aThread.get_vm()))
-            aThread.join()
-
-    def resetsshkey(self, vm, keypair, account=None, domainid=None):
-        """Resets SSH key"""
-        cmd = resetSSHKeyForVirtualMachine.resetSSHKeyForVirtualMachineCmd()
-        cmd.id = vm.id
-        cmd.keypair = keypair
-        cmd.account = account
-        cmd.domainid = domainid
-        return self.api_client.resetSSHKeyForVirtualMachine(cmd)
-
-    def update_sshkeypair(self, vm):
-        vm.stop(self.api_client)
-        self.resetsshkey(vm,
-                         self.keypair.name,
-                         account=self.account.user[0].account,
-                         domainid=self.account.domainid)
-        self.debug("Sshkey reset to - %s" % self.keypair.name)
-        vm.start(self.api_client)
-
-    def add_subnet_verify(self, network, services):
-        """verify required nic is present in the VM"""
-
-        self.debug("Going to add new ip range in shared network %s" %
-                   network.name)
-        cmd = createVlanIpRange.createVlanIpRangeCmd()
-        cmd.networkid = network.id
-        cmd.gateway = services["gateway"]
-        cmd.netmask = services["netmask"]
-        cmd.startip = services["startip"]
-        cmd.endip = services["endip"]
-        cmd.forVirtualNetwork = services["forvirtualnetwork"]
-        addedsubnet = self.api_client.createVlanIpRange(cmd)
-
-        self.debug("verify above iprange is successfully added in shared "
-                   "network %s or not" % network.name)
-
-        cmd1 = listVlanIpRanges.listVlanIpRangesCmd()
-        cmd1.networkid = network.id
-        cmd1.id = addedsubnet.vlan.id
-
-        allsubnets = self.api_client.listVlanIpRanges(cmd1)
-        self.assertEqual(
-            allsubnets[0].id,
-            addedsubnet.vlan.id,
-            "Check New subnet is successfully added to the shared Network"
-        )
-        return addedsubnet
 
     # =========================================================================
     # ---                    Gherkin style helper methods                   ---
@@ -916,7 +338,7 @@ class TestNuageConfigDrive(nuageTestCase):
                 metadata=True,
                 userdata=self.test_data[
                     "virtual_machine_userdata"]["userdata"],
-                sshkey=self.keypair.name)
+                ssh_key=self.keypair)
 
             expected_user_data1 = self.update_userdata(vm1, "helloworld vm1")
             self.verify_config_drive_content(vm1, public_ip_1,
@@ -929,7 +351,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(True),
                                              metadata=True,
                                              userdata=expected_user_data1,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             # After sshkey reset we need to have the vm password again
             vm1.password = vm1.resetPassword(self.api_client)
             self.debug("Password reset to - %s" % vm1.password)
@@ -944,7 +366,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm1.password),
                                              metadata=True,
                                              userdata=expected_user_data1,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             vm1.password = vm1.resetPassword(self.api_client)
             self.debug("Password reset to - %s" % vm1.password)
             self.debug("VM - %s password - %s !" %
@@ -955,7 +377,7 @@ class TestNuageConfigDrive(nuageTestCase):
             self.verify_config_drive_content(vm1, public_ip_1,
                                              self.PasswordTest(vm1.password),
                                              userdata=expected_user_data1,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("updating non-default nic as the default nic "
                        "of the multi-nic VM and enable staticnat...")
@@ -1040,7 +462,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm1.password),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("+++ Restarting the created Isolated network without "
                        "VR with cleanup...")
@@ -1051,7 +473,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm1.password),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("+++ Upgrade offering of created Isolated network with "
                        "a dns offering which spins a VR")
@@ -1149,14 +571,14 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm1.password),
                                              metadata=True,
                                              userdata=expected_user_data1,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("Updating userdata for VM - %s" % vm1.name)
             expected_user_data1 = self.update_userdata(vm1, "hello afterboot")
             self.verify_config_drive_content(vm1, public_ip_1,
                                              self.PasswordTest(vm1.password),
                                              userdata=expected_user_data1,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             self.debug("Resetting password for VM - %s" % vm1.name)
             self.reset_password(vm1)
             self.debug("SSHing into the VM for verifying its new password "
@@ -1171,7 +593,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm1.password),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("Updating userdata after migrating VM - %s" % vm1.name)
             expected_user_data1 = self.update_userdata(vm1,
@@ -1193,7 +615,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(False),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("Updating userdata for VM - %s" % vm1.name)
             expected_user_data1 = self.update_userdata(vm1,
@@ -1217,7 +639,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(False),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             self.update_provider_state("Disabled")
             expected_user_data1 = self.update_userdata(vm1,
                                                        "hello after recover")
@@ -1225,7 +647,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(False),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("+++ When template is not password enabled, "
                        "verify configdrive of VM - %s" % vm1.name)
@@ -1247,7 +669,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(False),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             vm1.delete(self.api_client, expunge=True)
             create_network1.network.delete(self.api_client)
 
@@ -1358,21 +780,21 @@ class TestNuageConfigDrive(nuageTestCase):
             self.verify_config_drive_content(vm, vpc_public_ip_1,
                                              self.PasswordTest(True),
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             expected_user_data = self.update_userdata(vm, "helloworld vm1")
             self.verify_config_drive_content(vm, vpc_public_ip_1,
                                              self.PasswordTest(True),
                                              metadata=True,
                                              userdata=expected_user_data,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("Resetting password for VM - %s" % vm.name)
             self.reset_password(vm)
             self.verify_config_drive_content(vm, vpc_public_ip_1,
                                              self.PasswordTest(vm.password),
                                              userdata=expected_user_data,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.generate_ssh_keys()
             self.update_sshkeypair(vm)
@@ -1380,7 +802,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(True),
                                              metadata=True,
                                              userdata=expected_user_data,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             # After sshkey reset we need to have the vm password again
             vm.password = vm.resetPassword(self.api_client)
             self.debug("Password reset to - %s" % vm.password)
@@ -1395,7 +817,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm.password),
                                              userdata=expected_user_data,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("Adding a non-default nic to the VM "
                        "making it a multi-nic VM...")
@@ -1405,7 +827,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm.password),
                                              metadata=True,
                                              userdata=expected_user_data,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             vm.password = vm.resetPassword(self.api_client)
             self.debug("Password reset to - %s" % vm.password)
             self.debug("VM - %s password - %s !" %
@@ -1415,7 +837,7 @@ class TestNuageConfigDrive(nuageTestCase):
             self.verify_config_drive_content(vm, vpc_public_ip_1,
                                              self.PasswordTest(vm.password),
                                              userdata=expected_user_data1,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("updating non-default nic as the default nic "
                        "of the multi-nic VM and enable staticnat...")
@@ -1476,7 +898,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm.password),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("+++ Restarting the created VPC Tier network without "
                        "cleanup...")
@@ -1487,7 +909,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm.password),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("+++ Restarting the created VPC Tier network with "
                        "cleanup...")
@@ -1498,7 +920,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm.password),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("Testing user data & password reset functionality "
                        " using configdrive in a VPC network with VR...")
@@ -1617,7 +1039,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm.password),
                                              metadata=True,
                                              userdata=expected_user_data1,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("Updating userdata for VM - %s" % vm.name)
             expected_user_data = self.update_userdata(vm,
@@ -1625,7 +1047,7 @@ class TestNuageConfigDrive(nuageTestCase):
             self.verify_config_drive_content(vm, vpc_public_ip_1,
                                              self.PasswordTest(vm.password),
                                              userdata=expected_user_data,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             self.debug("Resetting password for VM - %s" % vm.name)
             self.reset_password(vm)
             self.debug("SSHing into the VM for verifying its new password "
@@ -1640,7 +1062,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm.password),
                                              userdata=expected_user_data,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("Updating userdata after migrating VM - %s" % vm.name)
             expected_user_data = self.update_userdata(vm,
@@ -1648,7 +1070,7 @@ class TestNuageConfigDrive(nuageTestCase):
             self.verify_config_drive_content(vm, vpc_public_ip_1,
                                              self.PasswordTest(vm.password),
                                              userdata=expected_user_data,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             self.debug("Resetting password for VM - %s" % vm.name)
             self.reset_password(vm)
             self.debug("SSHing into the VM for verifying its new password "
@@ -1663,7 +1085,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(False),
                                              userdata=expected_user_data,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("Updating userdata for VM - %s" % vm.name)
             expected_user_data = self.update_userdata(vm,
@@ -1671,7 +1093,7 @@ class TestNuageConfigDrive(nuageTestCase):
             self.verify_config_drive_content(vm, vpc_public_ip_1,
                                              self.PasswordTest(False),
                                              userdata=expected_user_data,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             self.debug("Resetting password for VM - %s" % vm.name)
             self.reset_password(vm)
             self.debug("SSHing into the VM for verifying its new password "
@@ -1688,13 +1110,13 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(False),
                                              userdata=expected_user_data,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             self.update_provider_state("Disabled")
             self.verify_config_drive_content(vm, vpc_public_ip_1,
                                              self.PasswordTest(False),
                                              userdata=expected_user_data,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("+++ When template is not password enabled "
                        "verify configdrive of VM - %s" % vm.name)
@@ -1720,7 +1142,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(False),
                                              userdata=expected_user_data,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             vm.delete(self.api_client, expunge=True)
             create_tiernetwork.network.delete(self.api_client)
 
@@ -2024,7 +1446,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm1.password),
                                              metadata=True,
                                              userdata=expected_user_data,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             vm1.password = vm1.resetPassword(self.api_client)
             self.debug("Password reset to - %s" % vm1.password)
             self.debug("VM - %s password - %s !" %
@@ -2035,7 +1457,7 @@ class TestNuageConfigDrive(nuageTestCase):
             self.verify_config_drive_content(vm1, public_ip,
                                              self.PasswordTest(vm1.password),
                                              userdata=expected_user_data1,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("+++ Updating non-default nic as the default nic "
                        "of the multi-nic VM...")
@@ -2115,14 +1537,14 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm1.password),
                                              metadata=True,
                                              userdata=expected_user_data1,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("Updating userdata for VM - %s" % vm1.name)
             expected_user_data1 = self.update_userdata(vm1, "hello afterboot")
             self.verify_config_drive_content(vm1, public_ip,
                                              self.PasswordTest(vm1.password),
                                              userdata=expected_user_data1,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             self.debug("Resetting password for VM - %s" % vm1.name)
             self.reset_password(vm1)
             self.debug("SSHing into the VM for verifying its new password "
@@ -2137,7 +1559,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(vm1.password),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("Updating userdata after migrating VM - %s" % vm1.name)
             expected_user_data1 = self.update_userdata(vm1,
@@ -2159,7 +1581,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(False),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("Updating userdata for VM - %s" % vm1.name)
             expected_user_data1 = self.update_userdata(vm1,
@@ -2183,7 +1605,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(False),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             self.update_provider_state("Disabled")
             expected_user_data1 = self.update_userdata(vm1,
                                                        "hello after recover")
@@ -2191,7 +1613,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(False),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
 
             self.debug("+++ When template is not password enabled, "
                        "verify configdrive of VM - %s" % vm1.name)
@@ -2211,7 +1633,7 @@ class TestNuageConfigDrive(nuageTestCase):
                                              self.PasswordTest(False),
                                              userdata=expected_user_data1,
                                              metadata=True,
-                                             sshkey=self.keypair.name)
+                                             ssh_key=self.keypair)
             vm1.delete(self.api_client, expunge=True)
             shared_network.network.delete(self.api_client)
 
@@ -2272,7 +1694,7 @@ class TestNuageConfigDrive(nuageTestCase):
                 self.PasswordTest(vm1.password),
                 metadata=True,
                 userdata=expected_user_data,
-                sshkey=self.keypair.name)
+                ssh_key=self.keypair)
 
             for i in range(0, 300):
                 self.verify_config_drive_content(
@@ -2280,8 +1702,11 @@ class TestNuageConfigDrive(nuageTestCase):
                     self.PasswordTest(vm1.password),
                     metadata=True,
                     userdata=expected_user_data,
-                    sshkey=self.keypair.name,
+                    ssh_key=self.keypair,
                     ssh_client=ssh_client)
                 expected_user_data = \
                     self.update_userdata(vm1,
                                          'This is sample data %s' % i)
+
+if __name__ == "__main__" and __package__ is None:
+    __package__ = "integration.plugins.nuage"
